fix: add missing grid.svg asset

+++ backend/app/models/user.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, BigInteger
 from app.core.database import Base
 
 class User(Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True, index=True)
+    telegram_id = Column(BigInteger, unique=True, index=True)
     first_name = Column(String)
     username = Column(String, nullable=True)
     photo_url = Column(String, nullable=True)
